+++ pyproject.toml
@@ -7,6 +7,7 @@ requires-python = ">=3.12"
 dependencies = [
     "gradio>=5.29.0",
     "kokoro>=0.9.4",
+    "pip>=25.1.1",
 ]
 
 [dependency-groups]
@@ -19,5 +20,8 @@ dev = [
     "black>=25.1.0",
 ]
 
-[typos.default.extend-words]
+[tool.typos.default.extend-words]
 Chatacter = "Chatacter"
+
+[tool.pyrefly]
+python_interpreter = ".venv/Scripts/python"
