+++ src/voice_generator/__init__.py
@@ -5,16 +5,26 @@ Keeping this module tiny avoids pulling heavy dependencies into processes
 that only need the shared constants (e.g. docs builders, CI linters).
 """
 
-# `torch` is optional: gracefully fall back to CPU-only mode when it is absent.
-try:
-    import torch
+import os
+import random
+import gradio as gr
+import torch
+from gradio import Error
+from kokoro import KModel, KPipeline
 
-    CUDA_AVAILABLE: bool = torch.cuda.is_available()
-except ModuleNotFoundError:  # pragma: no cover – torch optional
-    torch = None  # type: ignore
-    CUDA_AVAILABLE = False
+CUDA_AVAILABLE: bool = torch.cuda.is_available()
 
 # Maximum allowed characters for TTS generation (`None` disables the cap).
-CHAR_LIMIT: int | None = 5000
+CHAR_LIMIT: int = 5000
 
-__all__ = ["CUDA_AVAILABLE", "CHAR_LIMIT"]
+models = {
+    gpu: KModel().to("cuda" if gpu else "cpu").eval()
+    for gpu in [False] + ([True] if CUDA_AVAILABLE else [])
+}
+pipelines = {
+    lang_code: KPipeline(lang_code=lang_code, model=False) for lang_code in "ab"
+}
+pipelines["a"].g2p.lexicon.golds["kokoro"] = "kˈOkəɹO"
+pipelines["b"].g2p.lexicon.golds["kokoro"] = "kˈQkəɹQ"
+
+__all__: list[str] = ["CUDA_AVAILABLE", "CHAR_LIMIT", "models", "pipelines"]
